# Improve type_caster for complex number types.