--- a/include/nanobind/stl/complex.h
+++ b/include/nanobind/stl/complex.h
@@ -16,54 +16,7 @@
 NAMESPACE_BEGIN(NB_NAMESPACE)
 NAMESPACE_BEGIN(detail)
 
-inline bool load_cmplx(PyObject *ob, uint8_t flags,
-                       std::complex<double> *out) noexcept {
-    bool is_complex = PyComplex_CheckExact(ob);
-#if !defined(Py_LIMITED_API)
-    if (is_complex || (flags & (uint8_t) cast_flags::convert)) {
-        Py_complex result = PyComplex_AsCComplex(ob);
-        if (result.real != -1.0 || !PyErr_Occurred()) {
-            *out = std::complex<double>(result.real, result.imag);
-            return true;
-        } else {
-            PyErr_Clear();
-        }
-    }
-#else
-#if Py_LIMITED_API < 0x030D0000
-    // Before version 3.13, __complex__() was not called by the Stable ABI
-    // functions PyComplex_{Real,Imag}AsDouble(), so we do so ourselves.
-    if (!is_complex && (flags & (uint8_t) cast_flags::convert)
-            && !PyType_IsSubtype(Py_TYPE(ob), &PyComplex_Type)
-            && PyObject_HasAttrString(ob, "__complex__")) {
-        PyObject* tmp = PyObject_CallFunctionObjArgs(
-                (PyObject*) &PyComplex_Type, ob, NULL);
-        if (tmp) {
-            double re = PyComplex_RealAsDouble(tmp);
-            double im = PyComplex_ImagAsDouble(tmp);
-            Py_DECREF(tmp);
-            if ((re != -1.0 && im != -1.0) || !PyErr_Occurred()) {
-                *out = std::complex<double>(re, im);
-                return true;
-            }
-        }
-        PyErr_Clear();
-        return false;
-    }
-#endif
-    if (is_complex || (flags & (uint8_t) cast_flags::convert)) {
-        double re = PyComplex_RealAsDouble(ob);
-        double im = PyComplex_ImagAsDouble(ob);
-        if ((re != -1.0 && im != -1.0) || !PyErr_Occurred()) {
-            *out = std::complex<double>(re, im);
-            return true;
-        } else {
-            PyErr_Clear();
-        }
-    }
-#endif
-    return false;
-}
+bool load_cmplx(PyObject*, uint8_t flags, std::complex<double> *out) noexcept;
 
 template <typename T> struct type_caster<std::complex<T>> {
     NB_TYPE_CASTER(std::complex<T>, const_name("complex"))