… tests/test_stl_ext.pyi.ref