Results for this PR were obtained with Python 3.12.8 and g++ 14.2.0.

Currently, after importing the module defined in `tests/test_stl.cpp` as `t` and defining
```
    class MyComplex:
        def __init__(self, real, imag):
            self.re = real
            self.im = imag
        def __complex__(self):
            return complex(self.re, self.im)
```
the following raises a TypeError:
```
    assert t.complex_value_double(MyComplex(1, 2)) == complex(1.0, 2.0)
```
Instead, it is expected that the [`__complex__`](https://docs.python.org/3.13/reference/datamodel.html#object.__complex__) method be called and the assertion be correct.
Note that the Stable ABI for [PyComplex_RealAsDouble,ImagAsDouble](https://docs.python.org/3.13/c-api/complex.html#c.PyComplex_RealAsDouble) was fixed in version 3.13.   See [cpython issue 109598](https://github.com/python/cpython/issues/109598)

Also, the current complex type_caster does not honor `nb::arg().noconvert()`.
The following test passes
```
    with pytest.raises(TypeError, match="incompatible function arguments"):
        t.complex_value_double_nc(0)
```
but given
```
    class MyInt(int):
        def __new__(cls, value):
            return super().__new__(cls, value)
```
the following tests `DID NOT RAISE <class 'TypeError'>` and fail:
```
    with pytest.raises(TypeError, match="incompatible function arguments"):
        t.complex_value_double_nc(MyInt(7))

    with pytest.raises(TypeError, match="incompatible function arguments"):
        t.complex_value_double_nc(0.0)

    with pytest.raises(TypeError, match="incompatible function arguments"):
        t.complex_value_float_nc(1.1 + 2.0j)  # Inexact narrowing conversion
```

This PR re-writes the complex type_caster.

Regarding performance, given the extension
```
    m.def("cadd", [](std::complex<double> x, std::complex<double> y) {
                          return x + y; });

    m.def("caddf", [](std::complex<float> x, std::complex<float> y) {
                          return x + y; });
```
and measuring time with
```
python3 -m timeit -n 100000000 -r 10 \
-s "import numpy as np, my_extension as me; x = complex(1); y = complex(2)" \
"me.cadd(x, y)"
```
I obtained the following results:

|                     | old 312 | new 312 | old abi3 | new abi3 |
|---------------------|---------|---------|----------|----------|
| cadd(complex)       | 27.6 ns | 23.7 ns | 31.0 ns  | 27.5 ns  |
| caddf(complex)      | 28.0 ns | 26.5 ns | 29.2 ns  | 29.8 ns  |
| cadd(np.complex128) | 42.8 ns | 30.0 ns | 45.7 ns  | 44.1 ns  |
| caddf(np.complex64) | 393. ns | 81.1 ns | 403. ns  | 284. ns  |

For the last two lines above, note that
```
>>> issubclass(np.complex128, complex)
True
>>> issubclass(np.complex64, complex)
False
```

I would expect that compiling `new abi3` with `-DPy_LIMITED_API=0x030D0000` would reduce `caddf(np.complex64)` by 2X, but I don't have Python 3.13 handy.