@hpkfft 2025-01-14
Sure.    Do you prefer not to include `<complex>` in the static library, or is it OK to do so for a single `cpp` file?