@wjakob 2025-01-14
It's fine to include in the cpp file.