@hpkfft 2025-01-15
I ran the `caddf(np.complex64)` benchmark with Python 3.13.1

|                     | (nsec) |
|---------------------|--------|
| old 313  non-STABLE | 380.   |
| new 313  non-STABLE | 82.1   |
| old abi3 0x030C0000 | 384.   |
| new abi3 0x030C0000 | 274.   |
| new abi3 0x030D0000 | 136.   |

The old complex caster used the C API functions from the stable abi regardless of whether `STABLE_ABI` was specified.
Prior to version 3.13, these functions had a bug (or, at least, some undesirable behavior) when the python object was neither `complex` nor a subclass of `complex`.   Since `np.complex64` is important to nanobind users, a workaround was needed.   The default `from_python</*Recursive=*/true>` made a temporary python object and called  `from_python</*Recursive=*/false>` on it.   It did this using the heuristic that an object with an attribute named `imag` should probably be converted to `complex`.  It's not foolproof, but it works for `np.complex64` .

The new complex caster uses `PyComplex_AsCComplex()`, which is not in the stable abi, when `STABLE_ABI` is not specified.   This is faster, and no workaround is needed.

The new caster uses the same complex C API functions as the old one did when `STABLE_ABI` is specified.   A workaround is applied when the object is not `complex`, not a subtype of `complex`, and has attribute `__complex__`.   Also, the workaround was made faster.  The function `from_python()` is no longer templated.

Also, templating is removed from `from_cpp()` since there seems to be no reason to distinguish rvalues from lvalues.

For non-STABLE_ABI builds, the new caster is much faster, and for STABLE_ABI builds it's faster.

As this [Note](https://docs.python.org/3/c-api/stable.html#stable-abi) mentions, the behavior of a stable abi function can change.   In particular, the workaround is no longer needed in version 3.13.   But, it is still perfectly correct to define Py_LIMITED_API=0x030C0000 and build with the workaround code in place.