@wjakob 2025-02-02
Actually, one question I realized after merging: is the `&& !PyType_IsSubtype(Py_TYPE(ob), &PyComplex_Type)` check really needed? I'm thinking that calling the `complex()` constructor should still do the right cast from a subtype.