@hpkfft 2025-02-03
Yes, the code would still be correct without that check, but it would be slower.   Referring to the table in my original comment for this PR, the case `cadd(np.complex128)` is 44ns for abi3 because Numpy's `complex128` is a subtype of Python's `complex`.  Without this check, that would increase to about 280ns.

The old code effectively had this check, because it used `PyComplex_Check()` to test whether to avoid calling the `complex()` constructor.   This check (after macro expansion) is found in Python's `object.h`:
```
static inline int PyObject_TypeCheck(PyObject *ob, PyTypeObject *type) {
    return Py_IS_TYPE(ob, type) || PyType_IsSubtype(Py_TYPE(ob), type);
}
```

The new code already has the boolean `is_complex` for an exact type match, so I used `PyType_IsSubtype()` directly to retain this optimization.  I think Numpy complex double precision, `np.complex128`, is an important enough use case to justify avoiding a performance regression.