@hpkfft 2025-06-05
This PR was primarily intended to avoid unintentional value-changing narrowing conversions (and the resulting loss of numerical precision).    Similar was done for real values.   See the nice table in  #829.
This intentionally affects overload resolution of functions.   See the changelog at the top (the "Files Changed" tab above).

I'm not sure what `python_obj` is exactly in your example above, or how it is bound using nanobind.

As a wild guess, I see the following:
https://github.com/apytypes/apytypes/blob/c99152b7221667d926e8353696cf55ec90f5d382/src/apycfixed_wrapper.cc#L371
and wonder if it should be `&APyCFixed::from_complex`.