@hpkfft 2025-06-08
@miklhh, @oscargus,
The test `nb::isinstance<std::complex<double>>()` tries the type caster without allowing conversions.
As you've observed, `nb::try_cast<std::complex<double>>()` allows all conversions (because the second parameter defaults to `bool convert = true`).

There's also https://nanobind.readthedocs.io/en/latest/api_core.html#_CPPv4N8nanobind10isinstanceE6handle6handle
This has nothing to do with the nanobind type caster.

Note that `numpy.complex128` is a subclass of python's `complex`.
Note that `numpy.complex64` is not.
```
>>> z = np.complex128(1+2j)
>>> isinstance(z, complex)
True
>>> z = np.complex64(1+2j)
>>> isinstance(z, complex)
False
```

The following, I think, will match python's complex or any subclass:

```
    m.def("iscomplex", [](const nb::object& obj) {
            return nb::isinstance(obj, nb::handle(&PyComplex_Type)); });
```